DSNet_hit_frame_prediction/requirements.txt
absl-py==1.4.0
brotlipy==0.7.0
cachetools==5.3.0
certifi
cffi
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h5py==3.8.0
idna
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy
oauthlib==3.2.2
opencv-python==4.7.0.72
ortools==9.6.2534
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.21.0
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==1.13.1+cu117
torchaudio==0.13.1
torchvision==0.14.1+cu117
tqdm==4.65.0
typing_extensions
tzdata==2023.3
ultralytics==8.0.88
urllib3
Werkzeug==2.2.3
win-inet-pton
win32-setctime==1.1.0
zipp==3.15.0
